refactor: extract useBillingData out of sidebar atom
Moving useBillingData (and the UsageData type) to
components/features/billing/use-billing-data.ts so use-run-actions
isn't taking a load-bearing dep on a sidebar UI atom. Four other
callers (outputs-section, plan-section, task panel, run actions)
move with it. usage-indicator.tsx keeps just the UI atoms
(UsageIndicator, TierBadge); its tests follow the same split.
