+++ happyhq/components/features/tasks/card/plan-section.tsx
@@ -1,7 +1,7 @@
 'use client'
 
+import { useBillingData } from '@/components/features/billing/use-billing-data'
 import { openChatSessionWindow } from '@/components/features/desktop/windows/chat/open-chat-window'
-import { useBillingData } from '@/components/features/sidebar/atoms/usage-indicator'
 import { ActivityHeader } from '@/components/features/tasks/atoms/activity-header'
 import { ConfirmRestartAlert } from '@/components/features/tasks/atoms/confirm-restart-alert'
 import { PlanFileRow } from '@/components/features/tasks/atoms/plan-file-row'
